Cater the changes in response to move logic in V1 module
