--- app/serializers/api/v3/npq_accounts_page_serializer.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-require "jsonapi/serializer/instrumentation"
-
-module Api
-  module V3
-    class NPQAccountsPageSerializer
-      include JSONAPI::Serializer
-      include JSONAPI::Serializer::Instrumentation
-
-      attributes :lead_provider_approval_status, :id
-    end
-  end
-end
